Programming/Movie Night/Resources/requirements.txt
pillow
qrcode
imageio
pyzbar
